test: remove Vitest in-source tests
