+++ .changeset/early-pianos-rhyme.md
@@ -1,0 +1,5 @@
+---
+'@makeswift/runtime': patch
+---
+
+Remove Vitest in-source tests.
