+++ packages/runtime/src/controls/link.test.ts
@@ -1,0 +1,44 @@
+import { describe, test, expect } from 'vitest'
+import { LinkControlData, copyLinkData } from './link'
+import { ReplacementContext } from '../state/react-page'
+
+describe.concurrent('link', () => {
+  test('page id is replaced by one in replacement context', () => {
+    // Arrange
+    const data: LinkControlData = {
+      type: 'OPEN_PAGE',
+      payload: {
+        pageId: 'UGFnZTpmNTdmMjQ2MS0wMGY3LTQzZWUtYmIwOS03ODdiNTUyYzUyYWQ=',
+        openInNewTab: false,
+      },
+    }
+    const expected = JSON.parse(
+      JSON.stringify(data).replace(
+        'UGFnZTpmNTdmMjQ2MS0wMGY3LTQzZWUtYmIwOS03ODdiNTUyYzUyYWQ=',
+        'testing',
+      ),
+    )
+
+    const replacementContext = {
+      elementHtmlIds: new Set(),
+      elementKeys: new Map(),
+      swatchIds: new Map(),
+      pageIds: new Map([['UGFnZTpmNTdmMjQ2MS0wMGY3LTQzZWUtYmIwOS03ODdiNTUyYzUyYWQ=', 'testing']]),
+      typographyIds: new Map(),
+      tableIds: new Map(),
+      tableColumnIds: new Map(),
+      fileIds: new Map(),
+      globalElementIds: new Map(),
+      globalElementData: new Map(),
+    }
+
+    // Act
+    const result = copyLinkData(data, {
+      replacementContext: replacementContext as ReplacementContext,
+      copyElement: node => node,
+    })
+
+    // Assert
+    expect(result).toMatchObject(expected)
+  })
+})
