+++ packages/runtime/src/controls/link.ts
@@ -1,4 +1,4 @@
-import { CopyContext, ReplacementContext } from '../state/react-page'
+import { CopyContext } from '../state/react-page'
 
 type OpenPageLink = {
   type: 'OPEN_PAGE'
@@ -66,49 +66,4 @@ export function copyLinkData(
   }
 
   return value
-}
-
-if (import.meta.vitest) {
-  const { describe, test, expect } = import.meta.vitest
-
-  describe.concurrent('link', () => {
-    test('page id is replaced by one in replacement context', () => {
-      // Arrange
-      const data: LinkControlData = {
-        type: 'OPEN_PAGE',
-        payload: {
-          pageId: 'UGFnZTpmNTdmMjQ2MS0wMGY3LTQzZWUtYmIwOS03ODdiNTUyYzUyYWQ=',
-          openInNewTab: false,
-        },
-      }
-      const expected = JSON.parse(
-        JSON.stringify(data).replace(
-          'UGFnZTpmNTdmMjQ2MS0wMGY3LTQzZWUtYmIwOS03ODdiNTUyYzUyYWQ=',
-          'testing',
-        ),
-      )
-
-      const replacementContext = {
-        elementHtmlIds: new Set(),
-        elementKeys: new Map(),
-        swatchIds: new Map(),
-        pageIds: new Map([['UGFnZTpmNTdmMjQ2MS0wMGY3LTQzZWUtYmIwOS03ODdiNTUyYzUyYWQ=', 'testing']]),
-        typographyIds: new Map(),
-        tableIds: new Map(),
-        tableColumnIds: new Map(),
-        fileIds: new Map(),
-        globalElementIds: new Map(),
-        globalElementData: new Map(),
-      }
-
-      // Act
-      const result = copyLinkData(data, {
-        replacementContext: replacementContext as ReplacementContext,
-        copyElement: node => node,
-      })
-
-      // Assert
-      expect(result).toMatchObject(expected)
-    })
-  })
 }
