+++ packages/runtime/src/prop-controllers/copy/image.ts
@@ -1,4 +1,4 @@
-import { CopyContext, ReplacementContext } from '../../state/react-page'
+import { CopyContext } from '../../state/react-page'
 import { ImageDescriptor, ImageValue } from '../descriptors'
 import { match, P } from 'ts-pattern'
 
@@ -14,37 +14,4 @@ export function copy(
       ([, v]) => context.replacementContext.fileIds.get(v.id) ?? v.id,
     )
     .otherwise(([, v]) => v)
-}
-
-if (import.meta.vitest) {
-  const { describe, test, expect } = import.meta.vitest
-
-  describe.concurrent('image copy', () => {
-    test('image is replaced by a one in replacement context', () => {
-      // Arrange
-      const data: ImageValue = 'file-id'
-      const expected = JSON.parse(JSON.stringify(data).replace('file-id', 'testing'))
-
-      const replacementContext = {
-        elementHtmlIds: new Set(),
-        elementKeys: new Map(),
-        swatchIds: new Map(),
-        fileIds: new Map([['file-id', 'testing']]),
-        typographyIds: new Map(),
-        tableIds: new Map(),
-        tableColumnIds: new Map(),
-        pageIds: new Map(),
-        globalElementIds: new Map(),
-        globalElementData: new Map(),
-      }
-
-      // Act
-      const result = copy({ type: 'Image', options: {} }, data, {
-        replacementContext: replacementContext as ReplacementContext,
-      })
-
-      // Assert
-      expect(result).toMatchObject(expected)
-    })
-  })
 }
