+++ packages/runtime/src/slate/BlockPlugin/__tests__/clearBlockKeyForDevice.test.tsx
@@ -1,6 +1,7 @@
 /** @jsxRuntime classic */
 /** @jsx jsx */
 
+import { describe, it, expect } from 'vitest'
 import { BlockActions } from '../'
 import { jsx, Paragraph, Text, Editor, Cursor } from '../../test-helpers'
