+++ packages/runtime/src/slate/BlockPlugin/__tests__/setBlockKeyForDevice.test.tsx
@@ -1,6 +1,7 @@
 /** @jsxRuntime classic */
 /** @jsx jsx */
 
+import { describe, it, expect } from 'vitest'
 import { BlockActions } from '..'
 import { DEFAULT_BREAKPOINTS } from '../../../state/modules/breakpoints'
 import { jsx, Paragraph, Text, Editor, Cursor, Anchor, Focus, Fragment } from '../../test-helpers'
