+++ packages/runtime/src/slate/BlockPlugin/__tests__/utils/getSelectedListItems.test.tsx
@@ -1,6 +1,7 @@
 /** @jsxRuntime classic */
 /** @jsx jsx */
 
+import { describe, it, expect } from 'vitest'
 import { getSelectedListItems } from '../../utils/getSelectedListItems'
 import {
   Editor,
