+++ packages/runtime/src/slate/InlineModePlugin/__tests__/inlineMode.test.tsx
@@ -1,6 +1,7 @@
 /** @jsxRuntime classic */
 /** @jsx jsxWithV2InlineEditor */
 
+import { describe, it, expect } from 'vitest'
 import { Editor as SlateEditor } from 'slate'
 import {
   jsxWithV2InlineEditor,
