+++ packages/runtime/src/slate/InlinePlugin/__tests__/getValue.test.tsx
@@ -1,6 +1,7 @@
 /** @jsxRuntime classic */
 /** @jsx jsx */
 
+import { describe, it, expect } from 'vitest'
 import { getValue } from '../getValue'
 import { InlineType } from '../..'
 import {
