+++ packages/runtime/src/slate/TypographyPlugin/__tests__/clearDeviceActiveTypography.test.tsx
@@ -1,6 +1,7 @@
 /** @jsxRuntime classic */
 /** @jsx jsx */
 
+import { describe, it, expect } from 'vitest'
 import { TypographyActions } from '../'
 import { jsx, Cursor, Editor, Paragraph, Text, Anchor, Focus } from '../../test-helpers'
