+++ packages/runtime/src/slate/TypographyPlugin/__tests__/normalization.test.tsx
@@ -1,6 +1,7 @@
 /** @jsxRuntime classic */
 /** @jsx jsx */
 
+import { describe, it, expect } from 'vitest'
 import { Editor as SlateEditor } from 'slate'
 import {
   Text,
