+++ packages/runtime/src/state/react-builder-preview.ts
@@ -35,8 +35,6 @@ import {
   registerMeasurable,
   registerPropControllers,
   registerPropControllersHandle,
-  registerDocument,
-  registerComponentHandle,
   unregisterBuilderComponent,
   unregisterMeasurable,
   unregisterPropControllers,
@@ -60,7 +58,7 @@ export type { Operation } from './modules/read-write-documents'
 export type { BoxModelHandle } from './modules/box-models'
 export { createBox, getBox, parse } from './modules/box-models'
 
-const reducer = combineReducers({
+export const reducer = combineReducers({
   documents: Documents.reducer,
   reactComponents: ReactComponents.reducer,
   boxModels: BoxModels.reducer,
@@ -614,7 +612,7 @@ function createAndRegisterPropControllers(
   }
 }
 
-function propControllerHandlesMiddleware(): Middleware<Dispatch, State, Dispatch> {
+export function propControllerHandlesMiddleware(): Middleware<Dispatch, State, Dispatch> {
   return ({ dispatch, getState }: MiddlewareAPI<Dispatch, State>) =>
     (next: ReduxDispatch<Action>) => {
       return (action: Action): Action => {
@@ -668,50 +666,6 @@ function propControllerHandlesMiddleware(): Middleware<Dispatch, State, Dispatch
         return next(action)
       }
     }
-}
-
-if (import.meta.vitest) {
-  const { describe, it, fn, expect } = import.meta.vitest
-
-  describe('propControllerHandlesMiddleware', () => {
-    it('registers prop controllers for element data', () => {
-      // Arrange
-      const documentKey = 'documentKey'
-      const element: ReactPage.Element = { key: 'elementKey', type: 'type', props: {} }
-      const store = createStore(reducer, applyMiddleware(thunk, propControllerHandlesMiddleware()))
-      const setPropControllers = fn()
-      const handle = new ElementImperativeHandle()
-
-      handle.callback(() => ({ setPropControllers }))
-
-      store.dispatch(registerDocument(ReactPage.createDocument(documentKey, element)))
-
-      // Act
-      store.dispatch(registerComponentHandle(documentKey, element.key, handle))
-
-      // Assert
-      expect(setPropControllers).toHaveBeenCalled()
-    })
-
-    it("doesn't register prop controllers for element references", () => {
-      // Arrange
-      const documentKey = 'documentKey'
-      const element: ReactPage.Element = { type: 'reference', key: 'elementKey', value: 'value' }
-      const store = createStore(reducer, applyMiddleware(thunk, propControllerHandlesMiddleware()))
-      const setPropControllers = fn()
-      const handle = new ElementImperativeHandle()
-
-      handle.callback(() => ({ setPropControllers }))
-
-      store.dispatch(registerDocument(ReactPage.createDocument(documentKey, element)))
-
-      // Act
-      store.dispatch(registerComponentHandle(documentKey, element.key, handle))
-
-      // Assert
-      expect(setPropControllers).not.toHaveBeenCalled()
-    })
-  })
 }
 
 function makeswiftApiClientSyncMiddleware(
