+++ packages/runtime/vite.test.config.ts
@@ -2,12 +2,4 @@
 
 import { defineConfig } from 'vite'
 
-export default defineConfig({
-  test: {
-    includeSource: ['src/**/*.{js,ts}'],
-    globals: true,
-  },
-  define: {
-    'import.meta.vitest': false,
-  },
-})
+export default defineConfig({})
